Added Kafka consumer and producer back
I'm not sure what happened but for some reason github wasn't letting my Kafka classes merge with main. It even said that main was fully up to date with it and it couldn't add anything. I just ended up reverting to and older commit and manually moving the files to main.