Made new Random Class
Had to make a new Random class that extends scala's random class. Because of a mistake I made at the start I had to roll back the scala version which led to scala's random class not having a between function. So I had to go in and remake the functions in my own Random class and use that instead