Further optimized data generation
created additional constructors for case classes, created an additional constructor for the row class that just takes the other case classes as arguments, turned the dataframes into datasets of their class, and turned the arrays of strings gained from the dataframes into arrays of their object instead. When printing all of the rows to the screen it finishes in 120 seconds, 40 seconds faster than before. If it's not being printed to the screen and just ran it finishes in 5 seconds.