Pulled data from Kafka
So 1. I overestimated how large their dataset would be and 2. possibly underestimated how good parquet's compression was. Their entire dataset is only around 37mb, so it can actually be uploaded to the branch. It'd still be better to pull from the topic when actually doing it but this helps as a form of backup/ incase it's too slow for someone else.